+++ CHANGELOG.md
@@ -38,6 +38,10 @@ _None_
 , i.e. `myfilter = "uppercase"` and then use it to invoke this filter with `{{ string|filter:myfilter }}`.  
   [Ilya Puchka](https://github.com/ilyapuchka)
   [#203](https://github.com/stencilproject/Stencil/pull/203)
+  
+- Added `map`, `compact` and `select` filters.  
+  [Ilya Puchka](https://github.com/ilyapuchka)
+  [#189](https://github.com/stencilproject/Stencil/pull/189)
 
 ### Deprecations
 
@@ -150,6 +154,7 @@ _None_
 - Improvements in error reporting.  
   [Ilya Puchka](https://github.com/ilyapuchka)
   [#167](https://github.com/stencilproject/Stencil/pull/167)
+
 
 ### Bug Fixes